Split runner project registration tests
